+++ compression/archive/tar/extract.go
@@ -24,7 +24,7 @@ import (
 // The extraction process is protected against zip-slip attacks, limited in terms
 // of file count, and file size.
 //
-//nolint:gocognit // This function is complex by nature
+//nolint:gocognit,gocyclo // This function is complex by nature
 func Extract(r io.Reader, outPath string, opts ...Option) error {
 	// Check arguments
 	if r == nil {
@@ -102,23 +102,18 @@ func Extract(r io.Reader, outPath string, opts ...Option) error {
 			return fmt.Errorf("tar entry %s is trying to escape the destination directory", hdr.Name)
 		}
 
-		// Determine if we need to skip the FS item
-		if dopts.OverwriteFilter != nil {
-			// Check existence
-			tfi, err := out.Stat(targetPath)
-			switch {
-			case err == nil:
-				// Nothing to do
-			case errors.Is(err, fs.ErrNotExist):
-				// Nothing to do
-			default:
-				return fmt.Errorf("unable to retrieve target file info: %w", err)
+		// Check existence
+		if fi, err := out.Lstat(targetPath); err == nil {
+			// Check if the file should be overwritten
+			if dopts.OverwriteFilter != nil && dopts.OverwriteFilter(targetPath, fi) {
+				continue
 			}
 
-			// Check if we have to skip the file overwrite
-			if tfi != nil && dopts.OverwriteFilter(targetPath, tfi) {
-				// Skip the file
-				continue
+			// Remove existing file
+			if !(fi.IsDir() && hdr.Typeflag == tar.TypeDir) {
+				if err := out.RemoveAll(targetPath); err != nil {
+					return fmt.Errorf("unable to remove existing file %q: %w", targetPath, err)
+				}
 			}
 		}
 
@@ -159,6 +154,33 @@ func Extract(r io.Reader, outPath string, opts ...Option) error {
 			if err := out.Chmod(targetPath, fs.FileMode(hdr.Mode)); err != nil {
 				return fmt.Errorf("unable to update file attributes: %w", err)
 			}
+
+			if dopts.RestoreTimes {
+				// Update file timestamps
+				if err := out.Chtimes(targetPath, hdr.AccessTime, hdr.ModTime); err != nil {
+					return fmt.Errorf("unable to update file timestamps %q: %w", targetPath, err)
+				}
+			}
+
+			//nolint:nestif // This has to adapt to the number of options
+			if dopts.RestoreOwner {
+				// Map UID/GID if needed
+				uid, gid := hdr.Uid, hdr.Gid
+				if dopts.UIDGIDMapper != nil {
+					uid, gid, err = dopts.UIDGIDMapper(hdr.Uid, hdr.Gid)
+					if err != nil {
+						return fmt.Errorf("unable to map UID/GID on file %q: %w", targetPath, err)
+					}
+					if uid < 0 || gid < 0 {
+						return fmt.Errorf("invalid UID/GID mapping on file %q: %w", targetPath, ErrAbortedOperation)
+					}
+				}
+
+				// Update file owner
+				if err := out.Chown(targetPath, uid, gid); err != nil {
+					return fmt.Errorf("unable to update file owner on %q: %w", targetPath, err)
+				}
+			}
 		default:
 			// ignore
 		}
@@ -167,7 +189,7 @@ func Extract(r io.Reader, outPath string, opts ...Option) error {
 	// Process symlinks and hardlinks
 	if len(symlinks) > 0 {
 		if err := processLinks(0, dopts.MaxSymlinkRecursion, outPath, out, symlinks); err != nil {
-			return fmt.Errorf("unable to process symlinks: %w", err)
+			return fmt.Errorf("unable to process links: %w", err)
 		}
 	}
 
@@ -200,8 +222,13 @@ func processLinks(level, maxRecursionDepth uint64, outPath string, out vfs.FileS
 		if filepath.IsAbs(hdr.Linkname) {
 			targetLinkName = filepath.Clean(hdr.Linkname)
 		} else {
-			//nolint:gosec // Path traversal is mitigated by the filesystem abstraction
-			targetLinkName = filepath.Join(filepath.Dir(targetName), hdr.Linkname)
+			if filepath.Dir(hdr.Linkname) == "." || strings.HasPrefix(hdr.Linkname, "..") {
+				//nolint:gosec // Path traversal is mitigated by the filesystem abstraction
+				targetLinkName = filepath.Join(filepath.Dir(targetName), hdr.Linkname)
+			} else {
+				// Make the relative link absolute
+				targetLinkName = "/" + filepath.Clean(hdr.Linkname)
+			}
 		}
 
 		// Check for zip-slip attacks (1st pass)
@@ -212,28 +239,23 @@ func processLinks(level, maxRecursionDepth uint64, outPath string, out vfs.FileS
 			return fmt.Errorf("tar entry %s is trying to escape the destination directory", hdr.Name)
 		}
 
-		// Check if the target link already exists, if not, add to next pass
-		if !out.Exists(targetLinkName) {
-			// Add to next pass
-			next = append(next, hdr)
-			continue
-		}
-
-		// Confirm filesystem membership
-		if _, _, err := out.Resolve(targetLinkName); err != nil {
-			return fmt.Errorf("unable to validate symlink target: %w", err)
-		}
-
 		switch hdr.Typeflag {
 		case tar.TypeLink:
 			// Create an absolute hardlink
 			if err := out.Link(targetLinkName, targetName); err != nil {
-				return fmt.Errorf("unable to create hardlink: %w", err)
+				return fmt.Errorf("unable to create hardlink %q to %q: %w", hdr.Linkname, hdr.Name, err)
 			}
 		case tar.TypeSymlink:
+			// Check if the target link already exists, if not, add to next pass
+			if !out.Exists(targetLinkName) {
+				// Add to next pass
+				next = append(next, hdr)
+				continue
+			}
+
 			// Create an absolute symlink
 			if err := out.Symlink(targetLinkName, targetName); err != nil {
-				return fmt.Errorf("unable to create symlink: %w", err)
+				return fmt.Errorf("unable to create symlink %q to %q: %w", hdr.Linkname, hdr.Name, err)
 			}
 		}
 	}