+++ vfs/api.go
@@ -7,6 +7,7 @@ import (
 	"io"
 	"io/fs"
 	"path/filepath"
+	"time"
 )
 
 const (
@@ -60,8 +61,12 @@ type FileSystem interface {
 	Exists(path string) bool
 	// Chmod changes the filemode of the given path.
 	Chmod(name string, mode fs.FileMode) error
+	// Chowwn changes the owner of the given path.
+	Chown(name string, uid, gid int) error
+	// Chtimes changes the access and modification times of the given path.
+	Chtimes(name string, atime, mtime time.Time) error
 	// Symlink creates a symbolink link.
-	Symlink(name, target string) error
+	Symlink(path, name string) error
 	// Link creates a hardlink.
 	Link(path, name string) error
 	// RemoveAll removes all path elements from the given path from the filesystem.