+++ vfs/chroot.go
@@ -8,6 +8,7 @@ import (
 	"io/fs"
 	"path/filepath"
 	"strings"
+	"time"
 )
 
 // Chroot returns a chrooted filesystem assuming an OS base filesystem as root
@@ -307,6 +308,38 @@ func (vfs chrootFS) Chmod(path string, mode fs.FileMode) error {
 	return vfs.unsafeFS.Chmod(path, mode)
 }
 
+// Chown delegates to the embedded unsafe FS after having confirmed the path
+// to be inside root. If the provided path violates this constraint, an error
+// of type ConstraintError is returned.
+//
+//nolint:wrapcheck // No need to wrap error
+func (vfs chrootFS) Chown(path string, uid, gid int) error {
+	// Apply root prefix
+	path = vfs.root.Join(path)
+
+	if err := isSecurePath(vfs.unsafeFS, vfs.root, path); err != nil {
+		return &ConstraintError{Op: "chown", Path: path, Err: err}
+	}
+
+	return vfs.unsafeFS.Chown(path, uid, gid)
+}
+
+// Chtimes delegates to the embedded unsafe FS after having confirmed the path
+// to be inside root. If the provided path violates this constraint, an error
+// of type ConstraintError is returned.
+//
+//nolint:wrapcheck // No need to wrap error
+func (vfs chrootFS) Chtimes(path string, atime, mtime time.Time) error {
+	// Apply root prefix
+	path = vfs.root.Join(path)
+
+	if err := isSecurePath(vfs.unsafeFS, vfs.root, path); err != nil {
+		return &ConstraintError{Op: "chtimes", Path: path, Err: err}
+	}
+
+	return vfs.unsafeFS.Chtimes(path, atime, mtime)
+}
+
 // Symlink delegates to the embedded unsafe FS after having confirmed the path
 // to be inside root. If the provided path violates this constraint, an error
 // of type ConstraintError is returned.
@@ -343,19 +376,19 @@ func (vfs chrootFS) Symlink(sourcePath, targetName string) error {
 // of type ConstraintError is returned.
 //
 //nolint:wrapcheck // No need to wrap error
-func (vfs chrootFS) Link(oldpath, newpath string) error {
+func (vfs chrootFS) Link(sourcePath, targetName string) error {
 	// Apply root prefix
-	oldpath = vfs.root.Join(oldpath)
-	newpath = vfs.root.Join(newpath)
+	sourcePath = vfs.root.Join(sourcePath)
+	targetName = vfs.root.Join(targetName)
 
-	if err := isSecurePath(vfs.unsafeFS, vfs.root, oldpath); err != nil {
-		return &ConstraintError{Op: "link", Path: oldpath, Err: err}
+	if err := isSecurePath(vfs.unsafeFS, vfs.root, sourcePath); err != nil {
+		return &ConstraintError{Op: "link", Path: sourcePath, Err: err}
 	}
-	if err := isSecurePath(vfs.unsafeFS, vfs.root, newpath); err != nil {
-		return &ConstraintError{Op: "link", Path: newpath, Err: err}
+	if err := isSecurePath(vfs.unsafeFS, vfs.root, targetName); err != nil {
+		return &ConstraintError{Op: "link", Path: targetName, Err: err}
 	}
 
-	return vfs.unsafeFS.Link(oldpath, newpath)
+	return vfs.unsafeFS.Link(sourcePath, targetName)
 }
 
 // Lstat delegates to the embedded unsafe FS after having confirmed the path