+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.0.4
 
-### 2024-02-14
+### Not released yet
 
 FEATURES: